feat(#15): added git alias
At first thought, I was going to wrap some of the Git commands in shell
functions to add default options. For example, when creating a branch with `git
switch -c branch` and no start-point was provided, I would use the default
branch. However, for now, I have decided not to pursue this because, first, it
doesn't make sense to have to edit in two places when adding an alias. Also,
the shell would need to "know" the alias. Second, I don't have that many
aliases.