feat(#18)!: using stow to manage symlink
I wanted to install everything with just one command, and having it as bash script would be too much work.

The main reason behid having this structure is because some files need to be in `$HOME` like `.zshenv` and `.gitconfig`

The only side effect of using stow was the organzation which I am not a fan, tried a couple of different organization
like: having `.config`  folder on the root, having a `stow-dir/.config` folder, also didn't wored, also I was not able
to use flag `--dotfiles` and not hide my folders bebause of this [bug](https://github.com/aspiers/stow/issues/33).