refactor!: using home-manager to manage dotfiles
It was decided to use the supervised and managed folders because the way that the home-manager handles files, the only
way to extend its generated configuration is by using extra parameters, for example `extraConfig`, so for the plugins
that I want home-manager to fully managed, like kitty because of the themes. For others I want just to create the
sybolic links on the .config folder, like nvim.